mnist_lab/pyproject.toml
[project]
name = "mnist-lab"
version = "0.1.0"
description = "Beginner-friendly PyTorch MNIST demo built with uv"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
  "matplotlib>=3.10.0",
  "pillow>=11.0.0",
  "torch>=2.8.0",
  "torchvision>=0.23.0",
  "tqdm>=4.67.0",
]

[project.scripts]
mnist-lab = "mnist_lab.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mnist_lab"]


[[tool.uv.index]]
url = 'https://mirrors.pku.edu.cn/pypi/web/simple'
default = true
